Reorganized application and switched to SQLAlchemy
A few things to note:

The forms should be using WTForms or something with better validation.

The pages themselves should be using ajax instead of jinja2 variables to fill selections.

I'm tempted to make an API-layer for updating the patrollers, locations, and activities. That
was a bit out of scope for this commit though.

to create database:
```
python manage.py createdb
```

to run app:
```
python manage.py runserver
```

**Note** The createdb does not handle already created databases. It creates a sqlite3 file with the tables.

One should look into flask-migrate to handle database migrations! Maybe add that to the roadmap.

Signed-off-by: Kevin <kevin@stealsyour.pw>